feat(public): fuse hooks for home
