+++ final-proyect/src/hooks/useHomeData.ts
@@ -9,14 +9,25 @@ type comment = {
     username: string;
   };
 };
+export const useFetchImg = () => {
+  const [productImgs, setProductImgs] = useState([]);
+  const randomNum = Math.floor(Math.random() * 100) + 1;
+
+  useEffect(() => {
+    fetch(`https://dummyjson.com/products/${randomNum}`)
+      .then((res) => res.json())
+      .then((datafetch) => setProductImgs(datafetch.images));
+  }, []);
+  return productImgs;
+};
 
 export const useHomeData = () => {
   const [comments, setComments] = useState<comment[]>();
   const about =
-    "Lorem ipsum dolor sit amet consectetur adipisicing elit. Nisi fugit recusandae minima? Laudantium quaerat veritatis, magnam doloremque nobis officia assumenda possimus explicabo tenetur at consequatur a ex eum hic nihil.";
+    "Lorem ipsum dolor sit amet consectetur adipisicing elit. Nisi fugit recusandae minima? Laudantium quaerat veritatis, magnam doloremque nobis officia assumenda possimus explicabo tenetur at consequatur a ex eum hic nihil.Lorem ipsum dolor sit amet consectetur adipisicing elit. Nisi fugit recusandae minima? Laudantium quaerat veritatis, magnam doloremque nobis officia assumenda possimus explicabo tenetur at consequatur a ex eum hic nihil.";
 
   useEffect(() => {
-    fetch("https://dummyjson.com/comments?limit=10")
+    fetch("https://dummyjson.com/comments?limit=14")
       .then((res) => res.json())
       .then((allcomments) => {
         console.log(allcomments);
